AndroidX converted, coroutines used for Async (probably not perfectly implemented but it'll do for now.)

+++ app/src/main/java/data/graveyard/factory/DefinitionFactory.kt
@@ -4,7 +4,7 @@ import data.room.entity.Definition
 import data.graveyard.DictionaryEntry
 
 class DefinitionFactory {
-    fun getDefinition(dictionaryEntry: DictionaryEntry, dictionaryID: Int, vocabularyID: Int): Definition {
+    fun getDefinition(dictionaryEntry: DictionaryEntry, dictionaryID: Long, vocabularyID: Long): Definition {
         return Definition(dictionaryEntry.definition,
                           dictionaryEntry.definitionLanguageCode,
                           dictionaryID,
